correcciones

+++ src/main/java/com/utn/space/venueaapi/Model/Notifications.java
@@ -16,7 +16,7 @@ import java.time.LocalDateTime;
 public class Notifications {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id_Notification;
+    private Long id_Notification;
 
     private LocalDateTime fecha;
     private String contenido;
