+++ src/main/java/com/utn/space/venueaapi/Repository/CommentsRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 
-public interface CommentsRepository extends JpaRepository<Comments,Integer> {
+public interface CommentsRepository extends JpaRepository<Comments,Long> {
 }
