+++ src/main/java/com/utn/space/venueaapi/Repository/NotificationsRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 
-public interface NotificationsRepository extends JpaRepository<Notifications,Integer> {
+public interface NotificationsRepository extends JpaRepository<Notifications,Long> {
 }
