+++ src/main/java/com/utn/space/venueaapi/Repository/ReservationRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 
-public interface ReservationRepository extends JpaRepository<Reservations,Integer> {
+public interface ReservationRepository extends JpaRepository<Reservations,Long> {
 }
